requirements.txt
openai
requests
beautifulsoup4
pandas
openpyxl
sqlalchemy
streamlit
sqlalchemy
google-generativeai
python-dotenv
